move bitstring to own mod
