+++ src/endpoint.rs
@@ -1,8 +1,8 @@
 //! # Endpoint
 //!
-//! `Endpoint` provides the entry point for DWN messages. Messages are routed
-//! to the appropriate handler for processing, returning a reply that can be
-//! serialized to a JSON object.
+//! `Endpoint` provides the entry point to the public API. Requests are routed
+//! to the appropriate handler for processing, returning a response that can 
+//! be serialized to a JSON object or directly to HTTP.
 
 use std::fmt::Debug;
 use std::ops::Deref;
